+++ pkg/processor/sumologicschemaprocessor/processor_test.go
@@ -39,7 +39,7 @@ func TestAddCloudNamespaceForLogs(t *testing.T) {
 			addCloudNamespace: true,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
 				return inputLogs
 			},
 			test: func(outputLogs plog.Logs) {
@@ -53,7 +53,7 @@ func TestAddCloudNamespaceForLogs(t *testing.T) {
 			addCloudNamespace: true,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ecs")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ecs")
 				return inputLogs
 			},
 			test: func(outputLogs plog.Logs) {
@@ -67,7 +67,7 @@ func TestAddCloudNamespaceForLogs(t *testing.T) {
 			addCloudNamespace: true,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_elastic_beanstalk")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_elastic_beanstalk")
 				return inputLogs
 			},
 			test: func(outputLogs plog.Logs) {
@@ -81,10 +81,10 @@ func TestAddCloudNamespaceForLogs(t *testing.T) {
 			addCloudNamespace: false,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_eks")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_lambda")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "azure_vm")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "gcp_app_engine")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_eks")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_lambda")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "azure_vm")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "gcp_app_engine")
 				return inputLogs
 			},
 			test: func(outputLogs plog.Logs) {
@@ -99,7 +99,7 @@ func TestAddCloudNamespaceForLogs(t *testing.T) {
 			addCloudNamespace: false,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
 				return inputLogs
 			},
 			test: func(outputLogs plog.Logs) {
@@ -113,10 +113,10 @@ func TestAddCloudNamespaceForLogs(t *testing.T) {
 			addCloudNamespace: true,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_lambda")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ecs")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_elastic_beanstalk")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_lambda")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ecs")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_elastic_beanstalk")
 				return inputLogs
 			},
 			test: func(outputLogs plog.Logs) {
@@ -166,7 +166,7 @@ func TestAddCloudNamespaceForMetrics(t *testing.T) {
 			addCloudNamespace: true,
 			createMetrics: func() pmetric.Metrics {
 				inputMetrics := pmetric.NewMetrics()
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
 				return inputMetrics
 			},
 			test: func(outputMetrics pmetric.Metrics) {
@@ -180,7 +180,7 @@ func TestAddCloudNamespaceForMetrics(t *testing.T) {
 			addCloudNamespace: true,
 			createMetrics: func() pmetric.Metrics {
 				inputMetrics := pmetric.NewMetrics()
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ecs")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ecs")
 				return inputMetrics
 			},
 			test: func(outputMetrics pmetric.Metrics) {
@@ -194,7 +194,7 @@ func TestAddCloudNamespaceForMetrics(t *testing.T) {
 			addCloudNamespace: true,
 			createMetrics: func() pmetric.Metrics {
 				inputMetrics := pmetric.NewMetrics()
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_elastic_beanstalk")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_elastic_beanstalk")
 				return inputMetrics
 			},
 			test: func(outputMetrics pmetric.Metrics) {
@@ -208,10 +208,10 @@ func TestAddCloudNamespaceForMetrics(t *testing.T) {
 			addCloudNamespace: false,
 			createMetrics: func() pmetric.Metrics {
 				inputMetrics := pmetric.NewMetrics()
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_eks")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_lambda")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "azure_vm")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "gcp_app_engine")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_eks")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_lambda")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "azure_vm")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "gcp_app_engine")
 				return inputMetrics
 			},
 			test: func(outputMetrics pmetric.Metrics) {
@@ -226,7 +226,7 @@ func TestAddCloudNamespaceForMetrics(t *testing.T) {
 			addCloudNamespace: false,
 			createMetrics: func() pmetric.Metrics {
 				inputMetrics := pmetric.NewMetrics()
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
 				return inputMetrics
 			},
 			test: func(outputMetrics pmetric.Metrics) {
@@ -240,10 +240,10 @@ func TestAddCloudNamespaceForMetrics(t *testing.T) {
 			addCloudNamespace: true,
 			createMetrics: func() pmetric.Metrics {
 				inputMetrics := pmetric.NewMetrics()
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_lambda")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ecs")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_elastic_beanstalk")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_lambda")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ecs")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_elastic_beanstalk")
 				return inputMetrics
 			},
 			test: func(outputMetrics pmetric.Metrics) {
@@ -293,7 +293,7 @@ func TestAddCloudNamespaceForTraces(t *testing.T) {
 			addCloudNamespace: true,
 			createTraces: func() ptrace.Traces {
 				inputTraces := ptrace.NewTraces()
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
 				return inputTraces
 			},
 			test: func(outputTraces ptrace.Traces) {
@@ -307,7 +307,7 @@ func TestAddCloudNamespaceForTraces(t *testing.T) {
 			addCloudNamespace: true,
 			createTraces: func() ptrace.Traces {
 				inputTraces := ptrace.NewTraces()
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ecs")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ecs")
 				return inputTraces
 			},
 			test: func(outputTraces ptrace.Traces) {
@@ -321,7 +321,7 @@ func TestAddCloudNamespaceForTraces(t *testing.T) {
 			addCloudNamespace: true,
 			createTraces: func() ptrace.Traces {
 				inputTraces := ptrace.NewTraces()
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_elastic_beanstalk")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_elastic_beanstalk")
 				return inputTraces
 			},
 			test: func(outputTraces ptrace.Traces) {
@@ -335,10 +335,10 @@ func TestAddCloudNamespaceForTraces(t *testing.T) {
 			addCloudNamespace: false,
 			createTraces: func() ptrace.Traces {
 				inputTraces := ptrace.NewTraces()
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_eks")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_lambda")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "azure_vm")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "gcp_app_engine")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_eks")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_lambda")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "azure_vm")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "gcp_app_engine")
 				return inputTraces
 			},
 			test: func(outputTraces ptrace.Traces) {
@@ -353,7 +353,7 @@ func TestAddCloudNamespaceForTraces(t *testing.T) {
 			addCloudNamespace: false,
 			createTraces: func() ptrace.Traces {
 				inputTraces := ptrace.NewTraces()
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
 				return inputTraces
 			},
 			test: func(outputTraces ptrace.Traces) {
@@ -367,10 +367,10 @@ func TestAddCloudNamespaceForTraces(t *testing.T) {
 			addCloudNamespace: true,
 			createTraces: func() ptrace.Traces {
 				inputTraces := ptrace.NewTraces()
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ec2")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_lambda")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_ecs")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.platform", "aws_elastic_beanstalk")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ec2")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_lambda")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_ecs")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.platform", "aws_elastic_beanstalk")
 				return inputTraces
 			},
 			test: func(outputTraces ptrace.Traces) {
@@ -420,8 +420,8 @@ func TestTranslateAttributesForLogs(t *testing.T) {
 			translateAttributes: true,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.account.id", "MyId1")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.account.id", "MyId2")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.account.id", "MyId1")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.account.id", "MyId2")
 
 				return inputLogs
 			},
@@ -440,8 +440,8 @@ func TestTranslateAttributesForLogs(t *testing.T) {
 			translateAttributes: false,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.account.id", "MyId1")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.account.id", "MyId2")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.account.id", "MyId1")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.account.id", "MyId2")
 
 				return inputLogs
 			},
@@ -460,8 +460,8 @@ func TestTranslateAttributesForLogs(t *testing.T) {
 			translateAttributes: true,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("not.actual.attr", "a1")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("maybe.an.attr", "a2")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("not.actual.attr", "a1")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("maybe.an.attr", "a2")
 				return inputLogs
 			},
 			test: func(outputLogs plog.Logs) {
@@ -479,9 +479,9 @@ func TestTranslateAttributesForLogs(t *testing.T) {
 			translateAttributes: true,
 			createLogs: func() plog.Logs {
 				inputLogs := plog.NewLogs()
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("cloud.account.id", "MyId")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("maybe.an.attr", "a2")
-				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutString("k8s.cluster.name", "A cool cluster")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("cloud.account.id", "MyId")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("maybe.an.attr", "a2")
+				inputLogs.ResourceLogs().AppendEmpty().Resource().Attributes().PutStr("k8s.cluster.name", "A cool cluster")
 
 				return inputLogs
 			},
@@ -529,9 +529,9 @@ func TestTranslateAttributesForMetrics(t *testing.T) {
 			translateAttributes: true,
 			createMetrics: func() pmetric.Metrics {
 				inputMetrics := pmetric.NewMetrics()
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("cloud.account.id", "MyId")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("maybe.an.attr", "a2")
-				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutString("k8s.cluster.name", "A cool cluster")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("cloud.account.id", "MyId")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("maybe.an.attr", "a2")
+				inputMetrics.ResourceMetrics().AppendEmpty().Resource().Attributes().PutStr("k8s.cluster.name", "A cool cluster")
 
 				return inputMetrics
 			},
@@ -580,9 +580,9 @@ func TestTranslateAttributesForTraces(t *testing.T) {
 			translateAttributes: true,
 			createTraces: func() ptrace.Traces {
 				inputTraces := ptrace.NewTraces()
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("cloud.account.id", "MyId")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("maybe.an.attr", "a2")
-				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutString("k8s.cluster.name", "A cool cluster")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("cloud.account.id", "MyId")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("maybe.an.attr", "a2")
+				inputTraces.ResourceSpans().AppendEmpty().Resource().Attributes().PutStr("k8s.cluster.name", "A cool cluster")
 
 				return inputTraces
 			},